✨ feat: implement group dialog for creating and editing groups with user search functionality
